fix:split FeatureCard
